Fix for browser.setValue being too fast for Elm separated into dedicated file
